chore: prune dead layout, add dev gates and repo map
Remove src/{ingestion,transcription,subtitle,translation} — leftover .pyc
directories from the pre-package layout with no Python sources. Untrack
.DS_Store and the test_auto_resume fixture; ignore .DS_Store going forward.

Make defaults.yaml the single source of default values: stage functions now
take model ids and merge thresholds as required parameters instead of
shadowing them with module-level DEFAULT_* constants. Replace the duplicated
TranslationSettings construction in cli.py and pipeline.py with
TranslationSettings.from_config, and the module-global model cache with an
lru_cache loader.

Add ruff (check only — the compact hand-wrapped style is deliberate, so no
auto-formatter), a Makefile for install/test/lint/check, and a README. Move
PRD/PIPELINE/EDGE_CASES into docs/ and rewrite CLAUDE.md as a working repo
map with commands, a per-module table, and the invariants that break silently.

+++ src/mlx_srt/srt.py
@@ -19,7 +19,7 @@ class Segment:
 
 
 def format_srt_time(seconds: float) -> str:
-    total_ms = max(0, int(round(seconds * 1000)))
+    total_ms = max(0, round(seconds * 1000))
     hours, remainder = divmod(total_ms, 3_600_000)
     minutes, remainder = divmod(remainder, 60_000)
     secs, millis = divmod(remainder, 1_000)
